refactor(tests): split checker dmarc and spf suites
Replace monolithic DMARC/SPF checker test modules with focused files, add shared checker test support helpers, and introduce a reusable RecordCheck factory used by CLI tests while preserving full test coverage.
