+++ tests/factories.py
@@ -2,7 +2,9 @@
 
 from __future__ import annotations
 
+from provider_check.checker import RecordCheck
 from provider_check.provider_config import ProviderConfig, ProviderVariable
+from provider_check.status import Status
 
 
 def make_provider_config(
@@ -39,4 +41,33 @@ def make_provider_config(
         variables=variables,
         short_description=short_description,
         long_description=long_description,
+    )
+
+
+def make_record_check(
+    *,
+    record_type: str = "MX",
+    status: Status = Status.PASS,
+    message: str = "ok",
+    details: dict[str, object] | None = None,
+    optional: bool = False,
+) -> RecordCheck:
+    """Build a record-check result object for tests.
+
+    Args:
+        record_type (str): DNS record type for the check.
+        status (Status): Status value for the record check.
+        message (str): Human-readable check message.
+        details (dict[str, object] | None): Optional details payload.
+        optional (bool): Whether the record check is optional.
+
+    Returns:
+        RecordCheck: Record-check model for assertions and mocked results.
+    """
+    return RecordCheck.with_status(
+        record_type,
+        status,
+        message,
+        details or {"found": ["mx"]},
+        optional=optional,
     )
